+++ src/service/cli/version.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const {version} = require('../../../../package.json')
+const {version} = require('../../../package.json')
 const chalk = require('chalk');
 
 module.exports  = {
